requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Faker==20.1.0
fastapi==0.104.1
h11==0.14.0
idna==3.6
iniconfig==2.0.0
mysql-connector-python==8.2.0
packaging==23.2
pluggy==1.3.0
protobuf==4.21.12
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
